Replaced locations of files and MakeFile
