+++ include/main.h
@@ -1,6 +1,6 @@
 //To be used in expenses.c
-extern double totalHours();        
-extern void convertDayAndHour();     
+extern double totalHours(int, int, int);        
+extern void convertDayAndHour(double);     
 
 //To be used in travelFees.c
 extern int getAmount();                          
@@ -17,6 +17,15 @@ extern double calculateAllowableExpenses();
 extern double registrationFee();
 extern void getMealFees();
 
+//To be used in milesDrivenLodgingParking.c
+extern double parkingFee(int);
+extern double hotelExpense(int);
+extern double milesDriven();
+
+//To be used in travelExpenseMain.c
+extern void setTravelValues();
+extern void setFareValues();
+extern void displayResults();
 
 //Variables 
 extern int numberOfDays; 
@@ -26,5 +35,17 @@ extern double testAirfare;
 extern double testCarRental;
 extern double testTaxi;
 extern double travelFeesTotal;
-//If these variables are global, we would not need setFareTravels
-//just need to add #include "../include/main.h" to other files to work
+extern double totalMiles;
+extern double totalHotelExpense;
+extern double totalParkingFees;
+extern double grandTotal;
+extern double hotelAllowance;
+extern double parkingAllowance;
+extern double totalMileDrivenAllowance;
+extern double totalAllowance;
+extern double registrationFees;
+extern double mealTotal;
+extern double totalReimbursement;
+extern double totalMilesDrivenAllowance;
+extern double totalAmountSaved;
+
