+++ src/travelExpenseMain.c
@@ -2,11 +2,36 @@
     This File is created by Andrew Valdez
 */
 
-#include "../include/main.h"
 #include <stdio.h>
 #include <stdlib.h>
+#include "../include/main.h"
 
-int main() //main should only hold values that must be inputted
+void displayResults(void);
+void setTravelValues(void);
+void setFareValues(void);
+
+int numberOfDays; 
+int departure, arrival;
+double hourTotal;
+double testAirfare;
+double testCarRental;
+double testTaxi;
+double travelFeesTotal;
+double totalMiles;
+double totalHotelExpense;
+double totalParkingFees;
+double grandTotal;
+double hotelAllowance;
+double parkingAllowance;
+double totalMileDrivenAllowance;
+double totalAllowance;
+double registrationFees;
+double mealTotal;
+double totalReimbursement;
+double totalMilesDrivenAllowance;
+double totalAmountSaved;
+
+int main(void) //main should only hold values that must be inputted
 {
     setTravelValues();
     setFareValues();
@@ -15,19 +40,7 @@ int main() //main should only hold values that must be inputted
     return EXIT_SUCCESS;
 }
 
-/*
-    display totals:
-
-    total days on trip
-    departure time
-    arrival time
-    total expenses 
-    total allowable expenses
-    total reimbursement
-    total amount saved
-*/
-
-void displayResults()
+void displayResults(void)
 {
     printf("------------------------\n");
     printf("Total Days & Hours Spent\n");
@@ -35,40 +48,48 @@ void displayResults()
     
     printf("Departure Time: %d\n", departure);
     printf("Arrival Time: %d\n", arrival);
-    printf("Total Days spent on the Business Trip: %d\n", numberOfDays);
     printf("Total Hours spent on the Business Trip: %.02f Hours\n", hourTotal);
     convertDayAndHour(hourTotal);
 
-    printf("\n\t============\n");
+    printf("\n\t================\n");
     printf("\tExpenses Summary\n");
-    printf("\t============\n");
+    printf("\t================\n");
     printf("Airfare   : $ %.2lf\n", testAirfare);
     printf("Car Rental: $ %.2lf\n", testCarRental);
     printf("Taxi Fares: $ %.2lf\n", testTaxi);
+    printf("Total Milages Driven: $ %.2lf\n", totalMiles);
+    printf("Total Lodging Fees: $ %.2lf\n", totalHotelExpense);
+    printf("Taxi Parking Fees: $ %.2lf\n", totalParkingFees);
+    
+    printf("TOTAL EXPENSES: %.2lf\n", grandTotal); 
+    printf("TOTAL ALLOWABLE : $ %.2lf\n", totalAllowance);
+    printf("TOTAL REIMBURSEMENT : $ %.2lf\n", totalReimbursement);
+    printf("SAVED AMOUNT : $ %.2lf\n", totalAmountSaved);
 
-    printf("Parking Fees:\n");
-    printf("Miles Driven:\n");
-
-    printf("Total Expenses: %.2lf\n", travelFeesTotal);
-    printf("Total Allowable expenses for the trip:\n");
-    printf("Excess that Must be Reimbursed:\n");
-    printf("Amount Saved if expense were under the total allowed:\n");
-
-/*
-    Should display total expense incurred by the person
-    total allowable expenses for the trip
-    the excess that must be reimbursed by the businessperson if any
-    the amount saved by the businessperson if the expenses were under the total allowed
-*/
+    //Incomplete NEED: meals eatern //must be tested
 }
 
-void setFareValues() {
+void setFareValues(void) {
     testAirfare = airfareAmount();
     testCarRental = carRental();
     testTaxi = taxiAmount();
+    totalMiles = milesDriven();
+    //registrationFees = registrationFee();
+    totalHotelExpense = hotelExpense(numberOfDays);
+    totalParkingFees = parkingFee(numberOfDays);
+    //mealTotal = getMealFees(numberOfDays, departure, arrival);
+
+    grandTotal = testAirfare + testCarRental + testTaxi + totalMiles + totalHotelExpense + totalParkingFees + registrationFees;
+
+    hotelAllowance = (double)numberOfDays * 90.00;
+    parkingAllowance = (double)numberOfDays * 6.00;
+    totalMileDrivenAllowance = totalMiles;
+    totalAllowance = hotelAllowance + parkingAllowance + totalMileDrivenAllowance + testAirfare + testCarRental;
+    totalReimbursement = grandTotal - totalAllowance;
+    totalAmountSaved = totalAllowance - grandTotal; //meal cost can be added through h file implementation
 }
 
-void setTravelValues() {
+void setTravelValues(void) {
     printf("\t==========================\n");
     printf("\tTRAVEL EXPENSE CALCULATION\n");
     printf("\t==========================\n");
