+++ src/travelFeesUnitTest.c
@@ -2,6 +2,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include "../include/main.h"
+
 int main()
 {
     printf("Unit Test using different arrays to represent different user input:\n");
